Move plotting funcion into plot
