+++ visualize_results.py
@@ -3,8 +3,6 @@ import click
 import numpy as np
 
 import matplotlib.pyplot as plt
-from matplotlib.widgets import CheckButtons
-
 
 from log_manager.log_reader import LogReader
 from stats import plot
@@ -23,34 +21,8 @@ def run(prefix):
     log_reader = LogReader(prefix=prefix)
     log_reader.read_data()
     log_reader.to_numeric(log_reader.header)
-    data_range = calc.calculate_data_range(log_reader, log_reader.header)
-    #print(data_range)
 
-    plt.subplots_adjust(right=0.8)
-
-    legend_info = []
-    graph_dict = {}
-    for group in log_reader.data:
-        legend_info.append(group)
-        line_i, = plt.plot(log_reader.data[group], label=group)
-        graph_dict[group] = line_i
-
-    plt.legend(legend_info, loc='upper left')
-    plt.ylabel("Average Elo Ranking")
-    plt.ylim(data_range[0], data_range[1])
-
-    def new_data(event):
-        """PEW."""
-        line = graph_dict[event]
-        line.set_visible(not line.get_visible())
-        plt.draw()
-
-    change_axis = plt.axes([0.85, 0.65, 0.13, 0.2])
-    change_buttons = CheckButtons(
-        change_axis, legend_info, [True]*len(legend_info))
-
-    change_buttons.on_clicked(new_data)
-    plt.show()
+    plot.plot_log_reader_data(log_reader)
 
 
 if __name__ == "__main__":
